implement move of the fruit and game logic

+++ src/Utility/DrawingUtility.java
@@ -50,7 +50,7 @@ public class DrawingUtility {
 		}
 	}
 	
-	public static void drawFruit(GraphicsContext gc, int x, int y) {
+	public static void drawFruit(GraphicsContext gc, double x, double y) {
 		gc.drawImage(fruit[(int) (Math.random()*10)], x, y, 50, 50);
 	}
